Separate gemini_llm_cost.py
